Using random temporary dir (but constant length)

+++ pack/poco/go.sh
@@ -8,7 +8,20 @@
 echo "Please run as ROOT (if needed): apt-get build-dep poco; apt-get install devscripts faketime"
 
 
-rm -rf mkdir /tmp/tmpbuild; mkdir -p /tmp/tmpbuild; chmod 700 /tmp/tmpbuild; cp libpcre-8.13.patch checksums /tmp/tmpbuild; cd /tmp/tmpbuild
+#/tmp/tmpbuild
+#/tmp/make1234
+#12345678
+
+dir_template="/tmp/makeXXXX"
+dir="$(mktemp -d "$dir_template" )"
+if [ -z "$dir" ] ; then echo "Problem creating temporary directory ($dir) from template ($dir_template)"; exit 1; fi
+echo $dir
+
+
+rm -rf "$dir"
+mkdir -p "$dir" ; chmod 700 "$dir";
+cp libpcre-8.13.patch checksums "$dir"
+cd "$dir" # <---------
 
 rm -rf build ; mkdir -p build ; cd build
 
@@ -27,8 +40,16 @@ do
 done
 
 echo "Checking sha512sum of builded libs"
-sha512sum out/usr/lib/*.so > checksums-to-verify
+sha512sum out/usr/lib/*.so > checksums-local
+cp checksums-local /tmp/ # XXX
+
+echo "TESTBASH:" # XXX
+bash
+
 
 echo "Differences:"
-diff checksums-to-verify ../checksums 
-echo "Builded packages are in: /tmp/tmpbuild/build. After checksum verification install with dpkg -i *.deb"
+diff -Nuar checksums-to-verify ../checksums 
+
+sha256deep -r "{$dir}/build/"
+
+echo "Builded packages are in: {$dir}/build. After checksum verification install with dpkg -i *.deb"
